🚚 Rename: 오타 수정
