+++ src/components/common/buttons/add-list-modal-button/index.tsx
@@ -9,7 +9,7 @@ import Drawer from "@/components/common/drawer";
 import Input from "@/components/common/forms/input";
 import Modal from "@/components/common/modal";
 import { useIsMobile, useToast, useToggle } from "@/hooks";
-import createTaskList from "@/lib/api/tast-list/create-tast-list";
+import createTaskList from "@/lib/api/task-list/create-task-list";
 import taskListAddEditSchema from "@/lib/schemas/task-list";
 import { IconPlusCurrent } from "@/public/assets/icons";
 import { TaskListAddEditInput } from "@/types/task-list";
